Allow building on an x86 host for testing.
For a number of business logic tests, it is possible to run the code on
x86 and still have enough coverage. Enabling devs to opt in to this
will allow use of standard tools such as Travis CI were esp32 hardware
or emulation may not be easily available.

This CL introduces a HOST_BUILD environment variable that components can
conditionally change their component.mk for buliding on a host system.
It updates components/mbedtls to use native sockets and avoid the esp32
accelerated math routines.

The CL also breaks out the library dependencies into a new
COMPONENT_LIBRARY_DEPS variable that custom makefiles in a project
can use to create host targets.